# speed up isequal and == for small tuples