--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -266,11 +266,19 @@ function _isequal(t1::Any16, t2::Any16)
     return true
 end
 
-function ==(t1::Tuple, t2::Tuple)
-    if length(t1) != length(t2)
+==(t1::Tuple, t2::Tuple) = (length(t1) == length(t2)) && _eq(t1, t2, false)
+_eq(t1::Tuple{}, t2::Tuple{}, anymissing) = anymissing ? missing : true
+function _eq(t1::Tuple, t2::Tuple, anymissing)
+    eq = t1[1] == t2[1]
+    if ismissing(eq)
+        return _eq(tail(t1), tail(t2), true)
+    elseif !eq
         return false
+    else
+        return _eq(tail(t1), tail(t2), anymissing)
     end
-    anymissing = false
+end
+function _eq(t1::Any16, t2::Any16, anymissing)
     for i = 1:length(t1)
         eq = (t1[i] == t2[i])
         if ismissing(eq)