Currently, `isequal` for inhomogeneous tuples allocates because it loops over the elements, so that the compiler cannot infer the type of the variables in the loop. This switches to a lisp style recursive definition for small tuples, in similar spirit to how `map` etc are implemented. The difference can be significant:

Before:
```julia
julia> @benchmark isequal((3,5.),(3,5))
BenchmarkTools.Trial: 
  memory estimate:  80 bytes
  allocs estimate:  3
  --------------
  minimum time:     46.539 ns (0.00% GC)
  median time:      51.482 ns (0.00% GC)
  mean time:        63.057 ns (12.80% GC)
  maximum time:     36.090 μs (99.81% GC)
  --------------
  samples:          10000
  evals/sample:     979
```
After:
```julia
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.649 ns (0.00% GC)
  median time:      1.896 ns (0.00% GC)
  mean time:        1.926 ns (0.00% GC)
  maximum time:     37.677 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```

I would like to do the same for `==`, but don't fully understand the current implementation with `missing`. It seems to me that, with the current implementation, if `missing` appears in the first slot of one or both of the two tuples (or all slots before its appearance are equal in both tuples) then `missing` is returned. However, if there are non-equal values before the `missing` is encountered, than `false` is returned. Is this the desired behaviour?