@nalimilan 2018-01-31
> I would like to do the same for ==, but don't fully understand the current implementation with missing. It seems to me that, with the current implementation, if missing appears in the first slot of one or both of the two tuples (or all slots before its appearance are equal in both tuples) then missing is returned. However, if there are non-equal values before the missing is encountered, than false is returned. Is this the desired behaviour?

Yes.