@Jutho 2018-01-31
Ok I understand the logic. Easy enough to implement :-).