@Jutho 2018-01-31
Before:
```julia
julia> @benchmark ==((3,5.),(3,5))
BenchmarkTools.Trial: 
  memory estimate:  80 bytes
  allocs estimate:  3
  --------------
  minimum time:     49.318 ns (0.00% GC)
  median time:      52.153 ns (0.00% GC)
  mean time:        65.059 ns (11.94% GC)
  maximum time:     34.785 μs (99.81% GC)
  --------------
  samples:          10000
  evals/sample:     979
```
After:
```julia
julia> @benchmark ==((3,5.),(3,5))
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.704 ns (0.00% GC)
  median time:      1.900 ns (0.00% GC)
  mean time:        2.059 ns (0.00% GC)
  maximum time:     36.608 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```