@nalimilan 2018-01-31
I just realized that the existing tests do not cover tuples with more than 16 elements, so it would be nice to add a few (including corner cases when missing values are present, see existing tests for examples).