@Jutho 2018-02-02
CI looks good; one question though: are the fallbacks for large tuples required for reduction functions like `==` and `isequal`, or only for `map`-like functions that produce tuple outputs. `hash` of a tuple does already seem to be a purely recursive definition irrespective of tuple length. I am only interested in relatively small tuples.

I could just benchmark to find out of course.