@JeffBezanson 2018-02-02
The fallbacks are usually still helpful due to code size issues.