@Jutho 2018-02-03
Ok then I think this PR is ready. Unless you want me to also add a large tuple fallback for `hash`.